Refactored the autoBindActivies method into a fancy new class "QuickTrack".
